Moves ApiResponse types to a new file and updates import paths for consistency
